模应用：uint8 改uint32

+++ interfaces/kits/napi/window_runtime/window_napi/js_window_utils.h
@@ -270,7 +270,7 @@ const std::map<WindowSizeChangeReason, RectChangeReason> JS_SIZE_CHANGE_REASON {
     { WindowSizeChangeReason::END,                   RectChangeReason::UNDEFINED  },
 };
 
-enum class ApiModalityType : uint8_t {
+enum class ApiModalityType : uint32_t {
     BEGIN = 0,
     WINDOW_MODALITY = BEGIN,
     APPLICATION_MODALITY,
